requirements.txt
icmplib
requests
psycopg2-binary
fastapi 
uvicorn
python-jose
